Include tests and runtime data writing in frontend
Refactor frontend to support different command verbs (incomplete)
