+++ src/devices/Arduino/samples/Arduino.sample.csproj
@@ -13,6 +13,7 @@
   <ItemGroup>
     <Compile Include="Arduino.sample.cs" />
     <Compile Include="RgbLedTest.cs" />
+    <Compile Include="TestCases.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\Bmxx80\Bmxx80.csproj" />
