+++ tools/ArduinoCsCompiler/Frontend/Frontend.csproj
@@ -22,6 +22,7 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\..\src\devices\Arduino\Arduino.csproj" />
+    <ProjectReference Include="..\..\..\src\devices\Arduino\samples\Arduino.sample.csproj" />
     <ProjectReference Include="..\ArduinoCsCompiler.csproj" />
   </ItemGroup>
 </Project>
